Exercises: Fix question creation errors after update from 1.11.x

+++ public/main/exercise/answer.class.php
@@ -812,15 +812,17 @@ class Answer
 
         if (count($this->position) > $this->new_nbrAnswers) {
             $i = $this->new_nbrAnswers + 1;
-            while ($this->position[$i]) {
-                $position = $this->position[$i];
-                $sql = "DELETE FROM $answerTable
+            if (isset($this->position[$i])) {
+                while ($this->position[$i]) {
+                    $position = $this->position[$i];
+                    $sql = "DELETE FROM $answerTable
                         WHERE
                             c_id = {$this->course_id} AND
                             question_id = '".$questionId."' AND
                             position ='$position'";
-                Database::query($sql);
-                $i++;
+                    Database::query($sql);
+                    $i++;
+                }
             }
         }
